Sending images in fp16; compile=False to make things great again on Nano

+++ inference_client.py
@@ -35,7 +35,7 @@ def create_zmq_socket(port):
 
 
 def send_and_evaluate_record(record, socket):
-  img = record[0].numpy()
+  img = record[0].numpy().astype('float16')
   val = record[1].numpy()
   socket.send(img)
   reply = socket.recv()
